arc-consistency enforcement algorithm
Basically, it’s a code implementation of the arc-consistency enforcement algorithm. We check the labels of each cell (object): if it has a label (state) in which it hasn't linked, then we remove this label from the cell. Then we check links of each cell(object): if a link between cells in the state(label), what was removed, exists, then we also remove this link. If we did some changes, the flag is set as 1, and the algorithm moves to a new iteration.